Updating verifications to use new connector code.  Remove token field from forms.  Store user email in session
