+++ src/views/water/licences-add/select-address.html
@@ -41,8 +41,6 @@
           <p class="lede"> Choose one of your licence addresses:</p>
         </legend>
 
-        <input type="hidden" name="token" value="{{ token }}" />
-
         <fieldset id="address">
 
         {{#each licences}}
