+++ src/views/water/licences-add/select-licences.html
@@ -30,7 +30,6 @@
       </h1>
 
       <form novalidate method="POST" action="/select-licences">
-        <input type="hidden" name="token" value="{{ token }}" />
 
         <fieldset id="licences">
           <legend>
